Blockchain formats: Cleanup code and support My Wallet v3
