+++ run/blockchain2john.py
@@ -29,15 +29,17 @@ if __name__ == '__main__':
             data = f.read()
             # try to detect the wallet format version, https://blockchain.info/wallet/wallet-format
             if b"guid" in data and args.json:  # v1
-                sys.stderr.write("My Wallet Version 1 seems to be used, try removing the --json option!\n")
-            if b"pbkdf2_iterations" in data and not args.json:  # v2
-                sys.stderr.write("My Wallet Version 2 seems to be used, try adding the --json option!\n")
+                sys.stderr.write("My Wallet Version 1 seems to be used, remove --json option!\n")
+                continue
+            if b"pbkdf2_iterations" in data and not args.json:  # v2/v3
+                sys.stderr.write("My Wallet Version 2 or 3 seems to be used, adding --json option is required!\n")
+                continue
 
             if args.json:
-                # hack for version 2.0 wallets
+                # hack for version 2.0 and 3.0 wallets
                 try:
                     decoded_data = json.loads(data.decode("utf-8"))
-                    if "version" in decoded_data and str(decoded_data["version"]) == "2":
+                    if "version" in decoded_data and (str(decoded_data["version"]) == "2" or str(decoded_data["version"]) == "3"):
                         payload = base64.b64decode(decoded_data["payload"])
                         iterations = decoded_data["pbkdf2_iterations"]
                         print("%s:$blockchain$v2$%s$%s$%s" % (
